refactor, limit logging from external modules
